requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
docker-pycreds==0.4.0
filelock==3.7.1
frozenlist==1.3.0
fsspec==2022.5.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.8.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
huggingface-hub==0.8.1
idna==3.3
joblib==1.1.0
Markdown==3.3.7
multidict==6.0.2
nltk==3.7
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
promise==2.3
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyDeprecate==0.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.4
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
sentencepiece==0.1.96
sentry-sdk==1.6.0
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
smmap==5.0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
torchmetrics==0.9.1
tqdm==4.64.0
transformers==4.20.1
typing_extensions==4.2.0
urllib3==1.26.9
wandb==0.12.19
Werkzeug==2.1.2
yarl==1.7.2
